refactor unit test increase unit test coverage to 100%

+++ tests/RealPathChangeDetectorTest.php
@@ -7,28 +7,50 @@ use Totoro1302\RealpathChangeDetector\RealPathChangeDetector;
 
 class RealPathChangeDetectorTest extends TestCase
 {
-    private const REAL_PATH_01 = __DIR__ . '/some/realPath01';
-    private const REAL_PATH_02 = __DIR__ . '/some/realPath02';
-    private const SYMLINK_PATH = __DIR__ . '/some/symlinkPath';
+    private const FIXTURES_PATH = __DIR__ . '/fixtures';
+    private const SYMLINK_PATH = __DIR__ . '/fixtures/symlinkDir';
+    private const DIR_BEFORE = __DIR__ . '/fixtures/dirBefore';
+    private const DIR_AFTER = __DIR__ . '/fixtures/dirAfter';
 
     public function testItReturnTrueIfPathChanged(): void
     {
-        mkdir(self::REAL_PATH_01, recursive: true);
-        symlink(self::REAL_PATH_01, self::SYMLINK_PATH);
+        $this->switchLinkTo(self::DIR_BEFORE);
 
-        $detector = new RealPathChangeDetector(self::SYMLINK_PATH);
+        $detector = new RealPathChangeDetector(self::SYMLINK_PATH . DIRECTORY_SEPARATOR . 'Source');
+
         $this->assertFalse($detector->pathHasChanged());
 
-        unlink(self::SYMLINK_PATH);
-        mkdir(self::REAL_PATH_02, recursive: true);
-        symlink(self::REAL_PATH_02, self::SYMLINK_PATH);
+        $this->switchLinkTo(self::DIR_AFTER);
 
         $this->assertTrue($detector->pathHasChanged());
+    }
 
-        unlink(self::SYMLINK_PATH);
-        rmdir(self::REAL_PATH_01);
-        rmdir(self::REAL_PATH_02);
+    public function testItThrowsExceptionIfOriginalPathDoesntExist(): void
+    {
+        $this->expectException(\RuntimeException::class);
+        new RealPathChangeDetector(self::FIXTURES_PATH . DIRECTORY_SEPARATOR . 'notExistingPath');
+    }
 
-        exec('rm -r ' . __DIR__ . '/some');
+    public function setUp(): void
+    {
+        $this->deleteSymlinkIfExists();
+    }
+
+    public function tearDown(): void
+    {
+        $this->deleteSymlinkIfExists();
+    }
+
+    private function deleteSymlinkIfExists(): void
+    {
+        if (file_exists(self::SYMLINK_PATH)) {
+            unlink(self::SYMLINK_PATH);
+        }
+    }
+
+    private function switchLinkTo(string $path): void
+    {
+        $this->deleteSymlinkIfExists();
+        symlink($path, self::SYMLINK_PATH);
     }
 }
